Huge, unsuccesful revision
I wanted to eliminate the nav layer and use physical arrow keys
instead. After some use, I don't see a reason why this would this
would be better. I enjoy the nav layer, so I am going to keep it.

Also put space of the left hand to avoid the space-shift same finger
bigram that comes from using cross thumb shift. This is also
relatively pointless. I think there are better ways to accomplish the
same thing. Also, it seems that many times sentences do not start with
letters on the righthand side of the Dvorak keyboard anyways.

The number row is throwing me for a loop, but I will keep it for a
little while because it will help me learn the symbol layer properly
and the sequential F keys that are paired with it are nice.

The new symbol layer has the potential to be good.

I will change the modifiers back to the old scheme, but with shift on
SPC and RET. These will the old inner thumb keys. The outer keys will
be tab and backspace and will have the nav layer and control on them
respectively. This mostly eliminates the Control problem. I can make
the backspace and enter on the nav layer be C-BKSPC and C-RET since
these don't need to modifier to actually be held like cycling with
C-SPC. The control on the thumb should allow me to reach all keys.